FIXED: Ran static analysis on most of projects, fixed logic errors, buffer overflow, unsigned integers checked if they were negative, misplaced parenthesis, minor performance issues, typos that caused assignments to the wrong variable.

+++ src/arcemu-world/Spell.h
@@ -1991,7 +1991,7 @@ class SERVER_DECL Spell : public EventableObject
 							this->Dur = 0;
 						c_dur = this->Dur;
 					}
-					if(sd->Duration1 >= 0 && !c_dur)
+					if((int32)sd->Duration1 >= 0 && !c_dur)
 					{
 						this->Dur = sd->Duration1;
 					}
